pyproject.toml
[tool.poetry]
name = "pahestor"
version = "0.1.0"
description = ""
authors = ["Tigran Saatchyan <mr.saatchyan@gmail.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.14"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"
redis = "^5.0.1"
celery = "^5.3.6"
flower = "^2.0.1"
django-filter = "^23.5"
setuptools = "^69.0.2"
django-jazzmin = "^2.6.0"


[tool.poetry.group.develop.dependencies]
ruff = "^0.1.7"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = ["D", "E", "PIE", "T20", "PYI", "Q", "RET", "SIM", "PTH"]
respect-gitignore = false
show-fixes = true
target-version = "py312"
exclude = [".venv", "migrations", "node_modules",]
preview = true
line-length = 79

[tool.ruff.lint.pycodestyle]
max-line-length = 79
max-doc-length = 72

[tool.ruff.lint]
select = [
    "D", # pydocstyle (D)
    "E", # pycodestyle (E, W)
    "PIE", # flake8-pie (PIE)
    "T20", # flake8-print (T20)
    "PYI", # flake8-pyi (PYI)
    "Q", # flake8-quotes (Q)
    "RET", # flake8-return (RET)
    "SIM", # flake8-simplify (SIM)
    "PTH", # flake8-use-pathlib (PTH)
    "TD", # flake8-todos (TD)
    "FIX", # flake8-fixmes (FIX)
    "FURB", # refurb (FURB)
    "RUF", # Ruff-specific rules (RUF)
    "C90", # mccabe (C90) (complex-structure)
    "I", # isort (I)
]
ignore = [
    "D417", # Missing argument descriptions in the docstring
    "D415", # First line should end with a period
    "D205", # 1 blank line required between summary line and description
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D106", # Missing docstring in public nested class
    "TD003", # Missing issue link on the line following this TODOs
    "FIX002", # todos found
    "FIX003", # XXX found
    "FIX004", # HACK found
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.black]
line-length = 79
target-version = ["py312"]
include = "\\.pyi?$"
exclude = "/(\\.venv|build|dist|migrations)/"
